pyproject.toml
[project]
name = "rag-search-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.60.0",
    "nltk==3.9.1",
    "numpy",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
]
